% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_Seurat.R
\name{dir10X}
\alias{dir10X}
\title{Create Seurat object}
\usage{
dir10X(
  seur = NULL,
  dat = NULL,
  minGenes = 500,
  regress = c("nCount_RNA"),
  species = "human"
)
}
\arguments{
\item{seur, }{a preexisting Seurat object}

\item{dat, }{a count Matrix (genes by cells)}

\item{minGenes, }{the min number of genes per cell}

\item{regress, }{things to regress out}

\item{species, }{the species being looked at}
}
\value{
A Seurat object
}
\description{
Creates a Seurat object given a count matrix or reprocesses an existing Seurat object
}
