% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDownsample.R
\name{dowsampleCurve}
\alias{dowsampleCurve}
\title{Downsample reads}
\usage{
dowsampleCurve(molinfo, pvals = 0.05 * 1:20)
}
\arguments{
\item{molinfo}{The molecular info from CellRanger}

\item{pvals}{The proportions to try downsampling to}
}
\value{
A list with one entry per proportion with the number of UMIs you get at that level of downsampling
}
\description{
Code to downsample reads to various levels and get number of UMI present at each level
}
