% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Azimuth.R
\name{GenerateReference}
\alias{GenerateReference}
\title{Generate a Azimuth Reference}
\usage{
GenerateReference(ref, meta.data)
}
\arguments{
\item{ref}{A seurat object}

\item{meta.data}{string representing a column in meta data}
}
\value{
A Azimuth reference
}
\description{
Turns a standard seurat obeject into a Azimuth reference, uses the column in mata.data as the column of interest
}
