% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep.for.MIMOSCA.R
\name{iteratePrep}
\alias{iteratePrep}
\title{Preps background to process with MIMOSCA}
\usage{
iteratePrep(seur, outdir, numIter = 100)
}
\arguments{
\item{seur}{The input Seurat object, see prepData for details}

\item{outdir}{The directory to output to}

\item{numIter}{the number of iterations to use}
}
\value{
NA
}
\description{
preps background input for MIMOSCA
}
