% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointAnalysis.R
\name{JointProcess}
\alias{JointProcess}
\title{Jointly process 2 Seurat objects}
\usage{
JointProcess(seur_ill, seur_ult)
}
\arguments{
\item{seur_ill, }{a Seurat objects to combine, with predicted.subclass columns in their meta data}

\item{seur_ult, }{a Seurat objects to combine, with predicted.subclass columns in their meta data}
}
\value{
A joint Seurat object
}
\description{
Processes 2 Seurat objects (one from Illumina, one from Ultima) together
}
