% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadPert.R
\name{LoadPertData}
\alias{LoadPertData}
\title{Load Perturb-seq}
\usage{
LoadPertData(dir)
}
\arguments{
\item{dir}{The CellRanger directory (out directory)}
}
\value{
seur Seurat object with Perturb-seq data
}
\description{
Loads CellRanger and DemuxEM results into Seurat for Pertub-seq
}
