% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeHTOMatrix.R
\name{makeHTOMat}
\alias{makeHTOMat}
\title{Create HTO Matrix}
\usage{
makeHTOMat(input, outfile)
}
\arguments{
\item{input}{The input 10X directory (the raw ouptut directory)}

\item{outfile}{The file the HTO matrix is to be written to}
}
\value{
Writes a HTO matrix ready for DemuxEM
}
\description{
Creates a HTO matrix to be used by DemuxEM using a CellRanger output with HTO data stored in the Custom field
}
