% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreNMF_Cells.R
\name{MakeTest_Cells}
\alias{MakeTest_Cells}
\title{Tests NMF}
\usage{
MakeTest_Cells(seur_ill, seur_ult, numTrain, numRep = 10, k = 15)
}
\arguments{
\item{seur_ill}{The first Seurat object}

\item{seur_ult}{The second Seurat object, on the same cells as the first}

\item{numTrain}{Number of cells to use for training}

\item{numRep}{The number of times repeated}

\item{k}{The number of factors to use}
}
\value{
dat A data frame of score information
}
\description{
Fits NMF on one seurat object and tests the accuracy on others using cell loadings
}
