% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep.for.MIMOSCA.R
\name{prepData}
\alias{prepData}
\title{Prep for MIMOSCA}
\usage{
prepData(seur, outfile)
}
\arguments{
\item{seur}{The input Seurat object. Needs Demux and Assign information from DemuxEM, feature_call to be perturbations in the mata dat}

\item{outfile}{The output file}
}
\value{
lst a listwhose first item is the matrix, the second the meta data
}
\description{
For a given Seurat object prepares input for the MIMOSCA tool
}
