% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalysis.R
\name{ProcessCellRangerOuput}
\alias{ProcessCellRangerOuput}
\title{Generates Seurat objects for comparision}
\usage{
ProcessCellRangerOuput(x, outdir)
}
\arguments{
\item{x}{a character array of length 5. The first entry is the sample name, the second the CellRanger directory of seq run 1, the third the CellRanger directory of seq run 2, the 4th the output of vireo on seq run 1 (N if not used), and the 5th the output of vireo on seq run 2 (N if not used)}

\item{outdir}{The directory to output results to}
}
\value{
lst List of images
}
\description{
Feed in two sequencing runs on same cells, outputs plots, etc. Note final plots in manuscript might be slightly different (modified later) but the underlying data should be the same.
}
