% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cNMF.R
\name{Repro_NMF}
\alias{Repro_NMF}
\title{Modified cNMF}
\usage{
Repro_NMF(
  seur,
  numSamp = 0,
  k = 15,
  numRep = 100,
  rho = 0.3,
  t = 0.03,
  getSeurat = F
)
}
\arguments{
\item{seur}{Seurat object}

\item{numSamp}{Number of cells to select (if <1 no downsampling occurs)}

\item{k}{The number of factors}

\item{numRep}{The number of times to run NMF internally before making the consensus}

\item{rho}{Internal parameter for cNMF}

\item{t}{The distance used for filtering out outliers}

\item{getSeruat}{If true returns Seurat object with cNMF added, else returns results as list}
}
\value{
out list of results (only returned if getSeurat=F)

seur Seurat object with cNMF added (only returned if getSeurat=T)
}
\description{
A modified version of the cNMF algorithm, implemented in R
}
