% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Azimuth.R
\name{RunAzimuth}
\alias{RunAzimuth}
\title{Run Azimuth with Reference}
\usage{
RunAzimuth(seur, reference_map)
}
\arguments{
\item{seur}{the Seurat object to label cell types for}

\item{reference_map}{Azimuth reference using the subclass column in the reference objects meta data}
}
\value{
Seurat object with Azimuth results and new assay for SCTransform
}
\description{
Applies azimuth to a Seurat object with a given reference reference
}
