% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcDEBetweenSeuratObjects.R
\name{runDE}
\alias{runDE}
\title{Compare DE between Seurat objects}
\usage{
runDE(seur1, seur2)
}
\arguments{
\item{seur1}{Seurat object number 1}

\item{seur2}{Seurat object number 2}
}
\value{
A dataframe with DE reslts with positive values being upregulated in seur2
}
\description{
Given two Seurat objects performs DE between them with Presto and reports the results
}
