% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunNebula.R
\name{RunNebula}
\alias{RunNebula}
\title{Run Nebula}
\usage{
RunNebula(seur, form_fixed, sampleCol, cpc = 0.1)
}
\arguments{
\item{seur}{the Seurat object to be tested}

\item{form_fixed}{The formula for the fixed effects (entries should be columns in the meta.data}

\item{sampleCol}{The column with the sample of origin information (HTO ID in our paper))}

\item{cpc}{Parameter used to filter out lowly expressed genes, see Nebula documentation}
}
\value{
The DE results in a format from Nebula
}
\description{
Runs Nebula on a Seurat object
}
