% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pert.DE.R
\name{TestPert}
\alias{TestPert}
\title{Perform DE analysis on 2 perturb-seq datasets}
\usage{
TestPert(
  seur1,
  seur2,
  ref_pert = "ONE_NON-GENE_SITE_18",
  pert = "feature_call"
)
}
\arguments{
\item{seur1}{The first Seurat obeject, should contain DemuxEM info in the Assign and Demux column}

\item{seur2}{The first Seurat obeject, should contain DemuxEM info in the Assign and Demux column}

\item{ref_pert}{The perturbation used as the reference in the DE}

\item{pert}{The columns in the meta.data with perturbation information}
}
\value{
ret A list of DE results for seur1 and seur2
}
\description{
Perform DE analysis on 2 perturb-seq datasets on the same set of cells (so different seq runs)
}
